[REFACTOR] fix merge conflicts and imports
