+++ src/transaksi_penjualan/test/enums/status_transaksi_test.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use crate::main::enums::status_transaksi::StatusTransaksi;
+    use crate::transaksi_penjualan::main::enums::status_transaksi::StatusTransaksi;
 
     #[test]
     fn test_status_transaksi_from_string() {
