+++ src/transaksi_penjualan/test/patterns/state_test.rs
@@ -3,9 +3,9 @@ mod tests {
     use chrono::Utc;
     use std::vec;
 
-    use crate::main::enums::status_transaksi::StatusTransaksi;
-    use crate::main::model::transaksi::{DetailProdukTransaksi, Transaksi};
-    use crate::main::patterns::state::TransaksiStateContext;
+    use crate::transaksi_penjualan::main::enums::status_transaksi::StatusTransaksi;
+    use crate::transaksi_penjualan::main::model::transaksi::{DetailProdukTransaksi, Transaksi};
+    use crate::transaksi_penjualan::main::patterns::state::TransaksiStateContext;
 
     #[test]
     fn test_masih_diproses_state_can_update() {
