+++ src/transaksi_penjualan/test/repository/transaksi_repository_test.rs
@@ -3,10 +3,10 @@ mod tests {
     use chrono::Utc;
     use uuid::Uuid;
 
-    use crate::main::model::transaksi::{DetailProdukTransaksi, Transaksi};
-    use crate::main::enums::status_transaksi::StatusTransaksi;
-    use crate::main::repository::transaksi_repository::TransaksiRepository;
-    use crate::main::repository::transaksi_repository_impl::TransaksiRepositoryImpl;
+    use crate::transaksi_penjualan::main::model::transaksi::{DetailProdukTransaksi, Transaksi};
+    use crate::transaksi_penjualan::main::enums::status_transaksi::StatusTransaksi;
+    use crate::transaksi_penjualan::main::repository::transaksi_repository::TransaksiRepository;
+    use crate::transaksi_penjualan::main::repository::transaksi_repository_impl::TransaksiRepositoryImpl;
 
     #[test]
     fn test_simpan_transaksi() {
